[utilze]코드리팩토링
