+++ front/src/App.js
@@ -1,7 +1,7 @@
 import Navbar from "./Navbar";
 import Navbar2 from "./Navbar2";
 import Carousel from "./Carousel";
-import MainContainer from "./MainContainer";
+import MainContainer from "./MainContainer/MainContainer";
 
 function App() {
   return (
