+++ front/src/MainContainer/MainContainer.jsx
@@ -1,0 +1,15 @@
+
+import React,{ useState} from "react";
+import DateCarousel from './DateCarousel';
+
+
+export default function MainContainer() {
+    const [dateBtnInfo, setDateBtnInfo] = useState({
+        btnIdx : 0,
+        total : 13,
+    });
+
+    return(
+       <DateCarousel dateBtnInfo={dateBtnInfo} setDateBtnInfo={setDateBtnInfo}></DateCarousel>
+    )
+}
